+++ executorWdl/src/main/scala/dx/executor/wdl/WdlWorkflowExecutor.scala
@@ -32,7 +32,7 @@ import wdlTools.types.WdlTypes._
 case class BlockIO(block: WdlBlock, logger: Logger)
 
 object WdlWorkflowExecutor {
-  def create(jobMeta: JobMeta): WdlWorkflowExecutor = {
+  def create(jobMeta: JobMeta, separateOutputs: Boolean): WdlWorkflowExecutor = {
     // parse the workflow source code to get the WDL document
     val (doc, typeAliases, versionSupport) =
       VersionSupport.fromSourceString(jobMeta.sourceCode, jobMeta.fileResolver)
@@ -42,7 +42,13 @@ object WdlWorkflowExecutor {
     val tasks = doc.elements.collect {
       case task: TAT.Task => task.name -> task
     }.toMap
-    WdlWorkflowExecutor(doc.source, workflow, versionSupport, tasks, typeAliases.toMap, jobMeta)
+    WdlWorkflowExecutor(doc.source,
+                        workflow,
+                        versionSupport,
+                        tasks,
+                        typeAliases.toMap,
+                        jobMeta,
+                        separateOutputs)
   }
 
   // this method is exposed for unit testing
@@ -87,8 +93,9 @@ case class WdlWorkflowExecutor(docSource: FileNode,
                                versionSupport: VersionSupport,
                                tasks: Map[String, TAT.Task],
                                wdlTypeAliases: Map[String, T_Struct],
-                               jobMeta: JobMeta)
-    extends WorkflowExecutor[WdlBlock](jobMeta) {
+                               jobMeta: JobMeta,
+                               separateOutputs: Boolean)
+    extends WorkflowExecutor[WdlBlock](jobMeta, separateOutputs) {
   private val logger = jobMeta.logger
   private lazy val evaluator = Eval(
       jobMeta.workerPaths,
@@ -323,7 +330,8 @@ case class WdlWorkflowExecutor(docSource: FileNode,
 
     private def launchCall(
         callInputs: Map[String, (T, V)],
-        nameDetail: Option[String] = None
+        nameDetail: Option[String] = None,
+        folder: Option[String] = None
     ): (DxExecution, ExecutableLink, String) = {
       logger.traceLimited(
           s"""|call = ${call}
@@ -382,13 +390,15 @@ case class WdlWorkflowExecutor(docSource: FileNode,
                   call.actualName,
                   callInputsIR,
                   nameDetail,
-                  instanceType.map(_.name))
+                  instanceType.map(_.name),
+                  folder = folder)
       (dxExecution, executableLink, execName)
     }
 
-    private def launchCall(): Map[String, ParameterLink] = {
+    private def launchCall(blockIndex: Int): Map[String, ParameterLink] = {
       val callInputs = evaluateCallInputs()
-      val (dxExecution, executableLink, callName) = launchCall(callInputs)
+      val (dxExecution, executableLink, callName) =
+        launchCall(callInputs, folder = Some(blockIndex.toString))
       jobMeta.createExecutionOutputLinks(dxExecution, executableLink.outputs, Some(callName))
     }
 
@@ -448,29 +458,6 @@ case class WdlWorkflowExecutor(docSource: FileNode,
       }
     }
 
-    /**
-      * A complex subblock requiring a fragment runner, or a subworkflow.
-      * For example:
-      *
-      *  if (flag) {
-      *    call zinc as inc3 { input: a = num}
-      *    call zinc as inc4 { input: a = num + 3 }
-      *
-      *    Int b = inc4.result + 14
-      *    call zinc as inc5 { input: a = b * 4 }
-      *  }
-      *
-      * There must be exactly one sub-workflow.
-      * @return
-      */
-    private def launchConditionalSubblock(): Map[String, ParameterLink] = {
-      assert(execLinkInfo.size == 1)
-      val executableLink = execLinkInfo.values.head
-      val callInputs = prepareSubworkflowInputs(executableLink)
-      val (dxExecution, _) = launchJob(executableLink, executableLink.name, callInputs)
-      jobMeta.createExecutionOutputLinks(dxExecution, executableLink.outputs)
-    }
-
     private def launchConditional(): Map[String, ParameterLink] = {
       val cond = block.target match {
         case Some(TAT.Conditional(expr, _, _)) =>
@@ -486,12 +473,31 @@ case class WdlWorkflowExecutor(docSource: FileNode,
           // }
           // The flag evaluates to true, so execute the inner call
           // and ensure it's output type is optional
-          launchCall()
+          launchCall(block.index)
         case (V_Boolean(true), BlockKind.ConditionalComplex) =>
-          // complex conditional block that requires a subworkflow
-          launchConditionalSubblock()
-        case (V_Boolean(false), _) =>
-          Map.empty
+          // A complex subblock requiring a fragment runner, or a subworkflow. For example:
+          //
+          // if (flag) {
+          //    call zinc as inc3 { input: a = num}
+          //    call zinc as inc4 { input: a = num + 3 }
+          //
+          //    Int b = inc4.result + 14
+          //    call zinc as inc5 { input: a = b * 4 }
+          // }
+          //
+          // There must be exactly one sub-workflow.
+          assert(execLinkInfo.size == 1)
+          val executableLink = execLinkInfo.values.head
+          val callInputs = prepareSubworkflowInputs(executableLink)
+          // there is no good human-readable name for a conditional, so we just
+          // use the current block index
+          val (dxExecution, _) =
+            launchJob(executableLink,
+                      executableLink.name,
+                      callInputs,
+                      folder = Some(block.index.toString))
+          jobMeta.createExecutionOutputLinks(dxExecution, executableLink.outputs)
+        case (V_Boolean(false), _) => Map.empty
         case _ =>
           throw new Exception(s"invalid conditional value ${cond}")
       }
@@ -967,7 +973,7 @@ case class WdlWorkflowExecutor(docSource: FileNode,
     override def launch(): Map[String, ParameterLink] = {
       val outputs: Map[String, ParameterLink] = block.kind match {
         case BlockKind.CallWithSubexpressions | BlockKind.CallFragment =>
-          launchCall()
+          launchCall(block.index)
         case BlockKind.ConditionalOneCall | BlockKind.ConditionalComplex =>
           launchConditional()
         case BlockKind.ScatterOneCall | BlockKind.ScatterComplex =>