+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## in dev
+
+* Adds `-separateOutputs` option to store output from each call in a separate folder
+
 ## 2.2.1 17-02-2021
 
 * Fix: native tasks now use instance type from native app/let, not wrapper task